fix(platform-api): cover llm_invocation in the startup schema check
